requirements.txt
flask
requests
beautifulsoup4
lxml
selenium
psycopg2-binary